Split InstanceViewPage into separate components
Pull out the components in InstanceViewPage.jsx into separate source
files.
